+++ controllers/SiteController.php
@@ -3,16 +3,12 @@
 namespace app\controllers;
 
 use Yii;
-// use yii\filters\AccessControl;
 use yii\web\Controller;
-// use yii\web\Response;
-// use yii\widgets\ActiveForm;
-// use app\models\LoginForm;
 
 class SiteController extends Controller
 {
     public $layout = 'landing';
-   
+
     /**
      * Displays homepage.
      *
