+++ controllers/TasksController.php
@@ -3,30 +3,14 @@
 namespace app\controllers;
 
 use Yii;
-use yii\web\Controller;
-use yii\filters\AccessControl;
 use app\models\Task;
 use app\models\Category;
 use app\models\Response;
 use app\models\FilterForm;
 use yii\web\NotFoundHttpException;
 
-class TasksController extends Controller
+class TasksController extends SecuredController
 {
-    public function behaviors()
-    {
-        return [
-            'access' => [
-                'class' => AccessControl::class,
-                'rules' => [
-                    [
-                        'allow' => true,
-                        'roles' => ['?'] // TODO Set roles
-                    ]
-                ]
-            ]
-        ];
-    }
 
     /**
      * Displays Tasks page.
